+++ strawberryfields/circuitspecs/circuit_specs.py
@@ -15,19 +15,24 @@
 Backend capabilities
 ====================
 
-**Module name:** :mod:`strawberryfields.devicespecs.device_specs`
+**Module name:** :mod:`strawberryfields.circuitspecs.circuit_specs`
 
-.. currentmodule:: strawberryfields.devicespecs.device_specs
+.. currentmodule:: strawberryfields.circuitspecs.circuit_specs
 
-The :class:`DeviceSpecs` class stores information about the capabilities of various execution backends.
-This information is used e.g., in :class:`Program` validation.
+The :class:`CircuitSpecs` class stores information about the capabilities of families or classes of circuits.
+For some circuit classes (e.g,, physical hardware chips), the specifications can be quite rigid. For other classes,
+e.g., circuit families supported by a particular simulator backend, the specifications can be more flexible and general.
+Key ingredients in a specification include: the primitive gates supported in the circuit class, the gates that can be
+decomposed to be supported, and the possible connectivity restrictions.
+
+This information is used e.g., in :class:`Program` validation and compilation.
 
 
 Classes
 -------
 
 .. autosummary::
-   DeviceSpecs
+   CircuitSpecs
 
 
 Code details
@@ -44,18 +49,18 @@ from blackbird.utils import to_DiGraph
 import strawberryfields.program_utils as pu
 
 
-class DeviceSpecs(abc.ABC):
-    """Abstract base class for describing execution backend capabilities."""
+class CircuitSpecs(abc.ABC):
+    """Abstract base class for describing circuit specifications."""
 
     short_name = ''
-    """str: short name of the device"""
+    """str: short name of the Circuit class"""
 
     @property
     @abc.abstractmethod
     def modes(self) -> Union[int, None]:
-        """The supported number of modes of the device.
+        """The number of modes supported by the circuit class.
 
-        If the device supports arbitrary number of modes, set this to 0.
+        If the circuit class supports arbitrary number of modes, set this to 0.
 
         Returns:
             int: number of supported modes
@@ -64,46 +69,46 @@ class DeviceSpecs(abc.ABC):
     @property
     @abc.abstractmethod
     def local(self) -> bool:
-        """Whether the backend supports local execution.
+        """Whether the circuit class can be executed locally (i.e., within a simulator).
 
         Returns:
-            bool: ``True`` if the backend supports local execution
+            bool: ``True`` if the circuit class supports local execution
         """
 
     @property
     @abc.abstractmethod
     def remote(self) -> bool:
-        """Whether the backend supports remote execution.
+        """Whether the circuit class supports remote execution.
 
         Returns:
-            bool: ``True`` if the backend supports remote execution
+            bool: ``True`` if the circuit class supports remote execution
         """
 
     @property
     @abc.abstractmethod
     def interactive(self) -> bool:
-        """Whether the backend can be used interactively, that is,
-        the backend state is not reset between engine executions.
+        """Whether the circuits in the class can be executed interactively, that is,
+        the registers in the circuit are not reset between engine executions.
 
         Returns:
-            bool: ``True`` if the backend supports interactive use
+            bool: ``True`` if the circuit supports interactive use
         """
 
     @property
     @abc.abstractmethod
     def primitives(self) -> Set[str]:
         """The primitive set of quantum operations directly supported
-        by the backend.
+        by the circuit class.
 
         Returns:
-            set[str]: the quantum primitives the backend supports
+            set[str]: the quantum primitives the circuit class supports
         """
 
     @property
     @abc.abstractmethod
     def decompositions(self) -> Dict[str, Dict]:
         """Quantum operations that are not quantum primitives for the
-        backend, but are supported via specified decompositions.
+        circuit class, but are supported via specified decompositions.
 
         This should be of the form
 
@@ -118,7 +123,7 @@ class DeviceSpecs(abc.ABC):
 
         Returns:
             dict[str, dict]: the quantum operations that are supported
-            by the backend via decomposition
+            by the circuit class via decomposition
         """
 
     @property
@@ -130,16 +135,16 @@ class DeviceSpecs(abc.ABC):
         Returns:
             dict[str, list]: a dictionary mapping an allowed quantum operation
             to a nested list of the form ``[[p0_min, p0_max], [p1_min, p0_max], ...]``.
-            where ``pi`` corresponds to the ``i`` th gate parameter.
+            where ``pi`` corresponds to the ``i`` th gate parameter
         """
         return dict()
 
     @property
     def graph(self):
-        """The allowed circuit topology of the backend device as a directed
+        """The allowed circuit topologies or connectivity of the class, modelled as a directed
         acyclic graph.
 
-        This property is optional; if arbitrary topologies are allowed by the device,
+        This property is optional; if arbitrary topologies are allowed in the circuit class,
         this will simply return ``None``.
 
         Returns:
@@ -166,12 +171,12 @@ class DeviceSpecs(abc.ABC):
 
     @property
     def circuit(self):
-        """The Blackbird circuit that will be accepted by the backend device.
+        """A rigid circuit template that defines this circuit specification.
 
-        This property is optional. If arbitrary topologies are allowed by the device,
+        This property is optional. If arbitrary topologies are allowed in the circuit class,
         **do not define this property**. In such a case, it will simply return ``None``.
 
-        If the device expects a specific template for the recieved Blackbird
+        If a backend device expects a specific template for the recieved Blackbird
         script, this method will return the serialized Blackbird circuit in string
         form.
 
@@ -181,7 +186,7 @@ class DeviceSpecs(abc.ABC):
         return None
 
     def compile(self, seq):
-        """Device-specific compilation method.
+        """Class-specific circuit compilation method.
 
         If additional compilation logic is required, child classes can redefine this method.
 
@@ -190,7 +195,7 @@ class DeviceSpecs(abc.ABC):
         Returns:
             List[Command]: modified circuit
         Raises:
-            CircuitError: the circuit is not valid for the device
+            CircuitError: the circuit is not valid within this circuit class
         """
         if self.graph is not None:
             # check topology
@@ -212,7 +217,7 @@ class DeviceSpecs(abc.ABC):
                 # TODO: try and compile the program to match the topology
                 # TODO: add support for parameter range matching/compilation
                 raise pu.CircuitError(
-                    "Program cannot be used with the device '{}' "
+                    "Program cannot be used with the CircuitSpec '{}' "
                     "due to incompatible topology.".format(self.short_name)
                 )