+++ strawberryfields/program.py
@@ -141,7 +141,7 @@ import warnings
 import networkx as nx
 
 import strawberryfields.circuitdrawer as sfcd
-import strawberryfields.devicespecs as specs
+import strawberryfields.circuitspecs as specs
 import strawberryfields.program_utils as pu
 from .program_utils import Command, RegRef, CircuitError, RegRefError
 
@@ -470,7 +470,7 @@ class Program:
         RegRef state remains consistent.
 
         Args:
-            target (str, DeviceSpecs): short name of the target circuit template, or the template itself
+            target (str, CircuitSpecs): short name of the target circuit template, or the template itself
 
         Keyword Args:
             optimize (bool): If True, try to optimize the program by merging and canceling gates.
@@ -481,7 +481,7 @@ class Program:
         Returns:
             Program: compiled program
         """
-        if isinstance(target, specs.DeviceSpecs):
+        if isinstance(target, specs.CircuitSpecs):
             db = target
             target = db.short_name
         elif target in specs.backend_specs:
@@ -544,7 +544,7 @@ class Program:
         if kwargs.get('optimize', False):
             seq = pu.optimize_circuit(seq)
 
-        # does the device have its own compilation method?
+        # does the circuit spec  have its own compilation method?
         if db.compile is not None:
             seq = db.compile(seq)